chore: Extract handle and buttons position logic to a custom hook
